requirements.txt
scikit-learn==1.3.2
pandas==2.1.4
nltk==3.8.1
Faker==22.0.0
imap-tools==1.5.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
schedule==1.2.1
joblib==1.3.2
